feat: add encryption to header columns

+++ model/header.go
@@ -1,5 +1,7 @@
 package model
 
+import "fmt"
+
 type Header struct {
 	ID    uint   `json:"id,omitempty" db:"id"`
 	Name  string `json:"name" db:"name"`
@@ -7,11 +9,11 @@ type Header struct {
 }
 
 func (c *Conn) SaveHeader(header *Header, userID, routeID uint) (err error) {
-	stmt := `
+	stmt := fmt.Sprintf(`
 		INSERT INTO	header(user_id, route_id, name, value)
-		VALUES ($1, $2, $3, $4)
+		VALUES ($1, $2, pgp_sym_encrypt($3, '%v'), pgp_sym_encrypt($4, '%v'))
 		RETURNING id
-	`
+	`, PG_CRYPT_KEY, PG_CRYPT_KEY)
 
 	row := c.db.QueryRow(stmt, userID, routeID, header.Name, header.Value)
 
@@ -24,11 +26,11 @@ func (c *Conn) SaveHeader(header *Header, userID, routeID uint) (err error) {
 }
 
 func (c *Conn) UpdateHeader(header Header, userID, routeID uint) (err error) {
-	stmt := `
+	stmt := fmt.Sprintf(`
 		UPDATE header
-		SET name = $1, value = $2
+		SET name = pgp_sym_encrypt($1, '%v'), value = pgp_sym_encrypt($2, '%v')
 		WHERE id= $3 AND user_id = $4 AND route_id = $5
-	`
+	`, PG_CRYPT_KEY, PG_CRYPT_KEY)
 
 	_, err = c.db.Exec(stmt, header.Name, header.Value, header.ID, userID, routeID)
 	if err != nil {
@@ -39,10 +41,10 @@ func (c *Conn) UpdateHeader(header Header, userID, routeID uint) (err error) {
 }
 
 func (c *Conn) DeleteHeader(headerName string, userID, routeID uint) (err error) {
-	stmt, err := c.db.Prepare(`
-		DELETE FROM header
-		WHERE "name" = $1 AND user_id = $2 AND route_id = $3
-	`)
+	stmt, err := c.db.Prepare(fmt.Sprintf(`
+		DELETE FROM header WHERE "name" = pgp_sym_encrypt($1, '%v') AND user_id = $2 AND route_id = $3
+		`, PG_CRYPT_KEY),
+	)
 
 	if err != nil {
 		return
@@ -57,11 +59,11 @@ func (c *Conn) DeleteHeader(headerName string, userID, routeID uint) (err error)
 }
 
 func (c *Conn) GetHeader(userID, routeID uint) (headers []Header, err error) {
-	stmt := `
-		SELECT id, name, value
+	stmt := fmt.Sprintf(`
+		SELECT id, pgp_sym_decrypt(name::bytea, '%v') as name, pgp_sym_decrypt(value::bytea, '%v') as value
 		FROM header
 		WHERE user_id = $1 AND route_id = $2
-	`
+	`, PG_CRYPT_KEY, PG_CRYPT_KEY)
 
 	rows, err := c.db.Query(stmt, userID, routeID)
 	if err != nil {
